rabin.c functions are moved in /libr/core
